# Taller Ninja - Desarrollo de un API/Paso 1/swagger-completo.yaml
swagger: "2.0"
info:
  description: "Ejemplo básico de API que permite la gestión mínima de una cuenta."
  version: "1.0.0"
  title: "Cuentas"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "cristobal.sepulveda@bbva.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/v2"
tags:
- name: "cuentas"
  description: "Gestión de una cuenta"
  externalDocs:
    description: "Más info"
    url: "http://swagger.io"
schemes:
- "http"
paths:
  /cuentas:
    post:
      tags:
      - "cuentas"
      summary: "Crear una cuenta nueva"
      description: ""
      operationId: "crearCuenta"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Objeto con toda la información necesaria para la creación de una cuenta."
        required: true
        schema:
          $ref: "#/definitions/Cuenta"
      responses:
        405:
          description: "Invalid input"
    put:
      tags:
      - "cuentas"
      summary: "Modificar una cuenta existente."
      description: ""
      operationId: "modificaCuenta"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Objeto con los datos de la cuenta a modificar."
        required: true
        schema:
          $ref: "#/definitions/Cuenta"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Cuenta no encontrada."
        405:
          description: "Validation exception"
  /cuentas/listado/{clienteId}:    
    get:
      tags:
      - "cuentas"
      summary: "Devolver las cuentas de un cliente"
      description: "Dado el identificador de un cliente, devuelve todas sus cuentas."
      operationId: "listaCuentas"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "clienteId"
        in: "path"
        description: "Id del cliente de la cuenta a consultar"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Cuenta"
        400:
          description: "Identificador de cuenta inválido"
          
  /cuentas/{cuentaId}:    
    get:
      tags:
      - "cuentas"
      summary: "Devolver los datos de una cuenta"
      description: "Dado el identificador de una cuenta de vuelve todos sus datos."
      operationId: "infoCuenta"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "cuentaId"
        in: "path"
        description: "Id de la cuenta a consultar"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Cuenta"
        400:
          description: "Identificador de cuenta inválido"
    delete:
      tags:
      - "cuentas"
      summary: "Eliminar una cuenta"
      description: ""
      operationId: "eliminarCuenta"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "api_key"
        in: "header"
        required: false
        type: "string"
      - name: "cuentaId"
        in: "path"
        description: "Id de la cuenta a eliminar"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "El id proporcionado es incorrecto."
        404:
          description: "Cuenta no encontrada."
      

definitions:
  Cuenta:
    type: "object"
    properties:
      cuentaId:
        type: "integer"
        format: "int64"
      cliente:
        type: "integer"
        format: "int64"
      saldo:
        type: "string"
        format: "int64"
      movimientos:
        type: "array"
        items:
          $ref: "#/definitions/Movimiento"
    xml:
      name: "Cuenta"
  Movimiento:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      fecha:
        type: "string"
        format: "date"
      saldo:
        type: "number"
        format: "double"
    xml:
      name: "Movimiento"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"